\name{VDSPCalibration-package}
\alias{VDSPCalibration-package}
\alias{VDSPCalibration}
\docType{package}
\title{
\packageTitle{VDSPCalibration}
}
\description{Implements statistical methods for designing and analyzing a calibration study}
\details{Implements statistical methods for design and analysis of a calibration study. The important functions are "samplesize": for sample size estimation; "sampletot": for sample selection, "calfun": for estimating calibrating equation and "chngpt": for estimating the piece-wise linear equation.
}
\author{Ramon Durazo-Arvizu,  Chris Sempos, and Lu Tian}
\references{Tian L., Durazo-Arvizu R. A., Myers G., Brooks S., Sarafin K., and Sempos C. T. (2014), The estimation of calibration equations for variables with heteroscedastic measurement errors, Statist. Med., 33, pages 4420-4436}
